+++ tencentcloud/resource_tc_gaap_layer4_listener_test.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/terraform"
+	sdkErrors "github.com/tencentcloud/tencentcloud-sdk-go/tencentcloud/common/errors"
 )
 
 func TestAccTencentCloudGaapLayer4Listener_basic(t *testing.T) {
@@ -291,6 +292,18 @@ func testAccCheckGaapLayer4ListenerDestroy(id *string, protocol string) resource
 		case "TCP":
 			listeners, err := service.DescribeTCPListeners(context.TODO(), nil, id, nil, nil)
 			if err != nil {
+				if sdkError, ok := err.(*sdkErrors.TencentCloudSDKError); ok {
+					switch sdkError.Code {
+					case GAAPResourceNotFound:
+						return nil
+
+					case "InvalidParameter":
+						if sdkError.Message == "ListenerId" {
+							return nil
+						}
+					}
+				}
+
 				return err
 			}
 			if len(listeners) > 0 {